requirements-uv.txt
# Managed by uv; prefer using `uv sync` with pyproject.toml
# This file lists runtime deps for quick review or ad-hoc installs

discord.py[voice]>=2.4,<2.6
discord-ext-voice-recv>=0.5.0
PyNaCl>=1.5
pyaudio>=0.2.14
aiohttp>=3.9
aiofiles>=23
groq>=0.13
gTTS>=2.5
yt-dlp>=2024.3
pillow>=10.0
imageio>=2.31
imageio-ffmpeg>=0.5
ffmpeg-python>=0.2
requests>=2.31
beautifulsoup4>=4.12
matplotlib>=3.8
seaborn>=0.13
pytz>=2023.3
playwright>=1.48
packaging>=24.0
torch>=2.7.0
diffusers>=0.31
transformers>=4.57.1
safetensors>=0.6.2
huggingface-hub>=0.34.0
accelerate>=1.11.0

